web-app/requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pymongo==4.12.0
pytest==8.3.5
pytest-flask==1.3.0
tomli==2.2.1
Werkzeug==3.1.3
zipp==3.21.0
Flask-PyMongo==3.0.1
python-dotenv==1.1.0